Initial version of culture-cron.
This should Just Work, I think.  I also set `/culture` in slack to point at it.
The secrets situation kind of sucks, but I think this approach sucks less than
the other options, and is roughly what we do for error-monitor-db, for example.

Test plan:
Deployed it, and did `/culture` in slack.  Also checked in the web
interface that the next scheduled culture times are what I expect.

Auditors: benjaminpollack, mroth